requirements.txt
flask-pymongo
flask
pydantic
PyYAML
openai
werkzeug